requirements.txt
Flask==2.2.5
Pillow==9.5.0
python-dotenv==1.0.0
google-generativeai==0.4.0
